POED-146  Budget: Spending By "District"
 Budget: Spending By "District"

Spending By "District" should be "Spending by Atolls and Islands". District is a term used in the software code. District in RMI are called "Atolls and Island", in FSM they are "States", in another country they could be something else. We use vocabs to make some general terms country specific. In the web app we use "filters" to generate the proper strict for those special vocabs.